Deleted the validators. Moved the validate logic to NativeEntity classes.
