+++ container_service_extension/rde/models/abstractNativeEntity.py
@@ -6,4 +6,14 @@ import abc
 
 
 class AbstractNativeEntity(abc.ABC):
-    pass
+
+    @abc.abstractmethod
+    def validate(self, request_spec: dict, current_spec: dict, operation: str):
+        """Validate the input_spec against current_spec.
+
+        :param dict request_spec: Request spec of the cluster
+        :param dict current_spec: Current status of the cluster
+        :param str operation: POST/PUT/DEL
+        :retur bool:
+        """
+        pass
