+++ container_service_extension/rde/models/rde_2_0_0.py
@@ -209,3 +209,13 @@ class NativeEntity:
         self.status = Status(**status) if isinstance(status, dict) else status
         self.kind = kind
         self.api_version = api_version
+
+    def validate(self, request_spec: dict, current_spec: dict, operation: str):
+        """Validate the input_spec against current_spec.
+
+        :param dict request_spec: Request spec of the cluster
+        :param dict current_spec: Current status of the cluster
+        :param str operation: POST/PUT/DEL
+        :retur bool:
+        """
+        raise NotImplementedError
